# GW-449-improve-query-performance-reading-sessions 